% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{ab_descriptives}
\alias{ab_descriptives}
\title{Attentional blink, descriptive stats}
\usage{
ab_descriptives(x, y)
}
\arguments{
\item{x}{Data frame}

\item{y}{Something}
}
\value{
Data frame
}
\description{
Attentional blink, descriptive stats.
}
