% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{ab_t_test}
\alias{ab_t_test}
\title{Attentional blink, t-tests}
\usage{
ab_t_test(x, y, lag1 = 3, lag2 = 8)
}
\arguments{
\item{x}{Data frame}

\item{y}{Something!}

\item{lag1}{Numeric}

\item{lag2}{Numeric}
}
\value{
Data frame
}
\description{
Attentional blink, t-tests.
}
