% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{age_summary_by_group}
\alias{age_summary_by_group}
\title{Age summary by group}
\usage{
age_summary_by_group(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame
}
\description{
Age summary by group.
}
