% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{aladins}
\alias{aladins}
\title{Returns values for Aladins Bayesian t-tests BF_t()}
\usage{
aladins(conditions, t_df, mean_diff_df)
}
\arguments{
\item{conditions}{Character vector}

\item{t_df}{Numeric}

\item{mean_diff_df}{Numeric}
}
\value{
Data frame
}
\description{
Returns values for Aladins Bayesian t-tests BF_t().
}
