% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{ant_descriptives}
\alias{ant_descriptives}
\title{ANT descriptives}
\usage{
ant_descriptives(scores)
}
\arguments{
\item{scores}{Data frame}
}
\value{
Data frame
}
\description{
ANT descriptives by score.
}
