% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{ant_descriptives_by_cue_type}
\alias{ant_descriptives_by_cue_type}
\title{ANT descriptives by cue type}
\usage{
ant_descriptives_by_cue_type(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame
}
\description{
ANT descriptives by cue type.
}
