% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{ant_plots}
\alias{ant_plots}
\title{ANT plots}
\usage{
ant_plots(df, condition, xlab)
}
\arguments{
\item{df}{Data frame}

\item{condition}{Character vector length 2}

\item{xlab}{String}
}
\value{
Data frame
}
\description{
ANT plots.
}
