% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{ant_scores}
\alias{ant_scores}
\title{Compute ANT scores for alerting, orienting and conflict}
\usage{
ant_scores(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame
}
\description{
Compute ANT scores for alerting, orienting and conflict.
}
