% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{apa_bayes_t}
\alias{apa_bayes_t}
\title{Report Bayesian t-test in APA format.}
\usage{
apa_bayes_t(bayes)
}
\arguments{
\item{bayes}{Bayesian t-test object}
}
\value{
String
}
\description{
Report Bayesian t-test in APA format.
}
