% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{blink}
\alias{blink}
\title{Calculate accuracy difference between lag 3 and lag 8}
\usage{
blink(rsvp_accuracy_3_8)
}
\arguments{
\item{rsvp_accuracy_3_8}{Data frame}
}
\value{
Data frame
}
\description{
Calculate accuracy difference between lag 3 and lag 8.
}
