% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{ci95_to_se}
\alias{ci95_to_se}
\title{Convert 95% confidence interval to standard error}
\usage{
ci95_to_se(u, l)
}
\arguments{
\item{u}{upper interval}

\item{l}{lower interval}
}
\value{
numeric
}
\description{
Convert 95% confidence interval to standard error
}
\details{
https://training.cochrane.org/handbook/current/chapter-06#section-6-3-1
}
