% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{collapse_meditation}
\alias{collapse_meditation}
\title{FAM+OMM vs. control}
\usage{
collapse_meditation(df, dir, prefix)
}
\arguments{
\item{df}{Data frame}

\item{dir}{String Directory}

\item{prefix}{String}
}
\value{
Data frame
}
\description{
FAM+OMM vs. control.
}
