% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{d}
\alias{d}
\title{Mean difference}
\usage{
d(m1, m2, sd)
}
\arguments{
\item{m1}{mean 1}

\item{m2}{mean 2}

\item{sd}{standard deviation}
}
\value{
numeric
}
\description{
Mean difference
}
\details{
Returns (m1 - m2) / sd. If sd is the pooled standard deviation, then this is Hedge's g.
}
