% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{d_ci95}
\alias{d_ci95}
\title{95% confidence interval for Cohen's d (Rosnow & Rosenthal, 2009)}
\usage{
d_ci95(d, n1, n2)
}
\arguments{
\item{d}{Cohen's d}

\item{n1}{n in group1}

\item{n2}{n in group2}
}
\value{
numeric
}
\description{
95% confidence interval for Cohen's d (Rosnow & Rosenthal, 2009)
}
\details{
95% confidence interval for Cohen's d (Rosnow & Rosenthal, 2009)
}
