% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{exclude_rt}
\alias{exclude_rt}
\title{Exclude RT outliers}
\usage{
exclude_rt(df, fast = 50, slow = 3)
}
\arguments{
\item{df}{data frame Must contain columns p, rt}

\item{fast}{numeric Fast RTs}

\item{slow}{numeric Standard deviation of slow RTs}
}
\value{
List of data frames
}
\description{
Exclude RT outliers.
}
\details{
Returns a list of data frames with keys, 'ok', 'too_fast', and 'too_slow',
which contain rows from df where the RT met the inclusion critera, was too
fast, or too slow respectively.
}
