% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{extract_rm_aov}
\alias{extract_rm_aov}
\title{Extract numbers from afex_aov object.}
\usage{
extract_rm_aov(afex_aov, effect = "condition")
}
\arguments{
\item{afex_aov}{afex_aov object}

\item{effect}{string}
}
\value{
tibble
}
\description{
Extract numbers from afex_aov object.
}
