% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{forest_bayes}
\alias{forest_bayes}
\title{Generate a forest plot from a brms model}
\usage{
forest_bayes(df, subgroup, var, xlab = "Standardised Mean Difference")
}
\arguments{
\item{df}{Data frame}

\item{subgroup}{String}

\item{var}{String}

\item{xlab}{String}
}
\value{
ggplot
}
\description{
Generate a forest plot from a brms model.
}
