% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{get_participants}
\alias{get_participants}
\title{Get participants}
\usage{
get_participants(df, c)
}
\arguments{
\item{df}{Data frame}

\item{c}{Condition}
}
\value{
List
}
\description{
Get participants.
}
