% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{process_demographics_survey}
\alias{process_demographics_survey}
\title{Process demographics survey}
\usage{
process_demographics_survey(survey, participants, by = "token", extra = F)
}
\arguments{
\item{survey}{Data frame containing demographics survey results pre-processed by \code{expfactory::process_expfactory_survey}}

\item{participants}{Data frame with columns p, condition, token}

\item{by}{String, column to join survey to participants (default='token')}

\item{extra}{Boolean, process extra demographics columns (default=F)}
}
\value{
Data frame
}
\description{
Process demographics survey.
}
