% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{process_posner}
\alias{process_posner}
\title{Process Posner task data files}
\usage{
process_posner(task_file, p)
}
\arguments{
\item{task_file}{String}

\item{p}{Numeric Participant number}
}
\value{
Data frame
}
\description{
Process Posner task data files.
}
