% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{report_aov}
\alias{report_aov}
\title{Report aov.}
\usage{
report_aov(aov, design = "mixed", effect = "1way")
}
\arguments{
\item{aov}{aov object}

\item{design}{string}

\item{effect}{string}
}
\value{
Character
}
\description{
Report aov.
}
