% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{report_collapsed_t}
\alias{report_collapsed_t}
\title{Report collapsed t.}
\usage{
report_collapsed_t(freq, bayes)
}
\arguments{
\item{freq}{Frequentist ANOVA object}

\item{bayes}{Bayesian ANOVA object}
}
\value{
String
}
\description{
Report collapsed t.
}
