% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{report_f}
\alias{report_f}
\title{Report f.}
\usage{
report_f(df_n, df_d, f, p)
}
\arguments{
\item{df_n}{Data frame Numerator degrees of freedom}

\item{df_d}{Data frame Denominator degrees of freedom}

\item{f}{Something f value}

\item{p}{String p value}
}
\value{
String
}
\description{
Report f.
}
