% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{report_t}
\alias{report_t}
\title{Report t.}
\usage{
report_t(df, results, c)
}
\arguments{
\item{df}{Data frame}

\item{results}{Something}

\item{c}{String condition}
}
\value{
String
}
\description{
Report t.
}
