% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{rsvp_accuracy}
\alias{rsvp_accuracy}
\title{RSVP Accuracy}
\usage{
rsvp_accuracy(rsvp, participants, combined = FALSE)
}
\arguments{
\item{rsvp}{Data frame}

\item{participants}{Data frame}

\item{combined}{Defaults to FALSE}
}
\value{
Data frame
}
\description{
RSVP Accuracy.
}
