% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{rsvp_by_group}
\alias{rsvp_by_group}
\title{Save frequentist and Baysian t-tests and Cohen's d for T1 and T2|T1}
\usage{
rsvp_by_group(df, group1, group2, dir, prefix)
}
\arguments{
\item{df}{Data frame}

\item{group1}{String}

\item{group2}{String}

\item{dir}{String Directory}

\item{prefix}{String}
}
\value{
Data frame
}
\description{
Save frequentist and Baysian t-tests and Cohen's d for T1 and T2|T1.
}
