% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{save_rds}
\alias{save_rds}
\title{Save data as RDS object.}
\usage{
save_rds(object, dir, filename)
}
\arguments{
\item{object}{R object}

\item{dir}{String Directory}

\item{filename}{String Filename}
}
\description{
Save data as RDS object.
}
