% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{sd_pooled}
\alias{sd_pooled}
\title{Pooled standard deviation}
\usage{
sd_pooled(n1, n2, sd1, sd2)
}
\arguments{
\item{n1}{n for group 1}

\item{n2}{n for group 2}

\item{sd1}{standard deviation for group 1}

\item{sd2}{standard deviation for group 2}
}
\value{
numeric
}
\description{
Pooled standard deviation
}
\details{
Returns the pooled standard deviation for two groups with same or different n.
}
