% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{spssSkewKurtosis}
\alias{spssSkewKurtosis}
\title{Skewness and kurtosis and their standard errors as implement by SPSS}
\usage{
spssSkewKurtosis(x)
}
\arguments{
\item{x}{Data frame}
}
\value{
Data frame
Reference: pp 451-452 of
http://support.spss.com/ProductsExt/SPSS/Documentation/Manuals/16.0/SPSS 16.0 Algorithms.pdf
See also: Suggestion for Using Powerful and Informative Tests of Normality,
Ralph B. D'Agostino, Albert Belanger, Ralph B. D'Agostino, Jr.,
The American Statistician, Vol. 44, No. 4 (Nov., 1990), pp. 316-321

This gives slightly different values for skew and kurtosis to
psych::describe(), which doesn't give SE of skew and kurtosis, which
you need to test for normality using a z-score
}
\description{
Skewness and kurtosis and their standard errors as implement by SPSS.
}
