% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{t1_accuracy}
\alias{t1_accuracy}
\title{T1 accuracy by participant}
\usage{
t1_accuracy(rsvp_accuracy)
}
\arguments{
\item{rsvp_accuracy}{Data frame}
}
\value{
Data frame
}
\description{
T1 accuracy by participant.
}
