% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{extract_aov}
\alias{extract_aov}
\title{Extract numbers from aov object.}
\usage{
extract_aov(aov, effect = "1way")
}
\arguments{
\item{aov}{aov object}
}
\value{
tibble
}
\description{
Extract numbers from aov object.
}
