% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{forest_bayes}
\alias{forest_bayes}
\title{Generate a forest plot from a brms model}
\usage{
forest_bayes(rem, xlab = "Standardised Mean Difference")
}
\arguments{
\item{rem}{brms model}

\item{xlab}{String}
}
\value{
ggplot
}
\description{
Generate a forest plot from a brms model.
}
