% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psy.phd.R
\name{set_effect}
\alias{set_effect}
\title{Set effect size for a study}
\usage{
set_effect(id, data, m1, m2, sd, n1, n2)
}
\arguments{
\item{id}{identifier for result}

\item{m1}{mean for group 1}

\item{m2}{mean for group 2}

\item{sd}{standard deviation}

\item{n1}{n in group 1}

\item{n2}{n in group 2}

\item{study}{data frame with publication column}
}
\value{
tibble
}
\description{
Set effect size for a study.
}
\details{
Calculates mean difference and 95% confidence interval.
}
